requirements-update.txt
flask-mail
email-validator
python-dotenv
Flask-WTF
WTForms
